Move benchmark run metrics projection inward
